Automatically backed up by Learn

+++ index.js
@@ -18,5 +18,5 @@ function dec(n) {
   return (n -= 1)
 }
 function makeInt(n) {
-  parseInt(n, 10)
+  parseFloat(n, 10)
 }
